# examples/openapi_action_schema.yaml
openapi: 3.1.0
info:
  title: FrontPocket API
  version: 0.2.0
  description: Source-backed memory retrieval API for FrontPocket.
servers:
  - url: http://localhost:8088
paths:
...
  /memory/session:
    post:
      operationId: upsertOrLoadSessionState
      summary: Create, update, or load session state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionRequest'
      responses:
        '200':
          description: Session state response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: deleteSessionState
      summary: Delete a session state record
      parameters:
        - $ref: '#/components/parameters/sessionIDQuery'
      responses:
        '200':
          description: Session delete response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-FrontPocket-Key
  parameters:
...
    sessionIDQuery:
      in: query
      name: session_id
      required: true
      schema:
        type: string
        minLength: 1
      description: Session identifier.
  responses:
    ValidationError:
      description: Invalid request payload.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    UnauthorizedError:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        detail:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
...
    SessionRequest:
      type: object
      required: [session_id]
      properties:
        session_id:
          type: string
          minLength: 1
        project:
          type: string
        active_summary:
          type: string
        recent_memory_ids:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties:
            type: string
        ttl_seconds:
          type: integer
          minimum: 1
        load_only:
          type: boolean
    SessionState:
      type: object
      required: [session_id, updated_at]
      properties:
        session_id:
          type: string
        project:
          type: string
        active_summary:
          type: string
        recent_memory_ids:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties:
            type: string
        updated_at:
          type: string
          format: date-time
    SessionResponse:
      type: object
      required: [found]
      properties:
        found:
          type: boolean
        state:
          $ref: '#/components/schemas/SessionState'